Switch everything to doubles

+++ matrix.cpp
@@ -1,10 +1,10 @@
 #include "matrix.hpp"
 
 #include <stdexcept>
-matrix::matrix(int rows, int cols) : data(rows, std::vector<int>(cols)) {}
+matrix::matrix(int rows, int cols) : data(rows, std::vector<double>(cols)) {}
 
 std::ostream& operator<<(std::ostream& out, matrix& mat) {
-    for(std::vector<int> row : mat.data) {
+    for(std::vector<double> row : mat.data) {
         for(int item : row) {
             out << item << " ";
         }
@@ -20,7 +20,7 @@ matrix dot(matrix& a, matrix& b) {
     matrix result(a.height(), b.width());
     for(int i = 0; i < a.height(); ++i) {
         for(int j = 0; j < b.width(); ++j) {
-            int sum = 0;
+            double sum = 0;
             for(int k = 0; k < a.width(); ++k) {
                 sum += a.data[i][k] * b.data[k][j];
             }
@@ -30,15 +30,15 @@ matrix dot(matrix& a, matrix& b) {
     return result;
 }
 
-std::vector<int>& matrix::operator[](int index) {
+std::vector<double>& matrix::operator[](int index) {
     return data[index];
 }
 
-int matrix::get(int row, int col) {
+double matrix::get(int row, int col) {
     return data[row][col];
 }
 
-void matrix::set(int row, int col, int num) {
+void matrix::set(int row, int col, double num) {
     data[row][col] = num;
 }
